Re #20438. Refactor Chunker to not depend on HDF5.
Step towards better testability.
